feat: move to monorepo setup and rename core package to @navettejs/core"
BREAKING CHANGE: main package renamed to @navettejs/core
