+++ .github/workflows/pull-request.yml
@@ -21,7 +21,7 @@ jobs:
       - name: Install dependencies
         run: yarn install
       - name: Test
-        run: yarn build
+        run: yarn turbo build
   test:
     name: Unit tests
     runs-on: ubuntu-latest
@@ -40,7 +40,7 @@ jobs:
       - name: Install dependencies
         run: yarn install
       - name: Test
-        run: yarn test
+        run: yarn turbo test
   mutation:
     name: Mutations testing
     runs-on: ubuntu-latest
@@ -59,4 +59,4 @@ jobs:
       - name: Install dependencies
         run: yarn install
       - name: Test
-        run: yarn test:mutation
+        run: yarn turbo mutation-test
