+++ .gitignore
@@ -38,3 +38,6 @@ coverage
 # stryker temp files
 .stryker-tmp
 reports
+
+# Turborepo
+.turbo
