+++ .releaserc.yml
@@ -1,5 +1,5 @@
 plugins:
   - '@semantic-release/commit-analyzer'
   - '@semantic-release/release-notes-generator'
-  - '@semantic-release/npm'
+  - 'semantic-release-yarn'
   - '@semantic-release/github'
