+++ packages/core/assets/logo.png
