+++ packages/core/assets/logo_whitebg.png
