+++ packages/core/assets/small_logo_whitebg.png
