+++ packages/core/jest-stryker.config.ts
