+++ packages/core/jest.config.ts
