+++ packages/core/package.json
@@ -1,0 +1,62 @@
+{
+  "name": "@navettejs/core",
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "check-types": "tsc --noEmit",
+    "build": "vite build",
+    "test": "tsc --project tsconfig.test.json && jest",
+    "mutation-test": "stryker run"
+  },
+  "files": [
+    "./assets",
+    "./dist",
+    "./README.md",
+    "./package.json"
+  ],
+  "main": "./dist/navette.umd.js",
+  "module": "./dist/navette.es.js",
+  "exports": {
+    ".": {
+      "import": "./dist/navette.es.js",
+      "require": "./dist/navette.umd.js"
+    }
+  },
+  "types": "./dist/main.d.ts",
+  "devDependencies": {
+    "@stryker-mutator/core": "^8.7.1",
+    "@stryker-mutator/jest-runner": "^8.7.1",
+    "@stryker-mutator/typescript-checker": "^8.7.1",
+    "@types/jest": "^29.5.14",
+    "@types/jsdom": "^21.1.7",
+    "@types/node": "^22.14.1",
+    "globals": "^16.0.0",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "pinst": "^3.0.0",
+    "ts-jest": "^29.3.2",
+    "ts-node": "^10.9.2",
+    "type-testing": "^0.2.0",
+    "typescript": "~5.7.2",
+    "vite": "^6.2.0",
+    "vite-plugin-cp": "^6.0.0",
+    "vite-plugin-dts": "4.5.0"
+  },
+  "packageManager": "yarn@4.9.1+sha512.f95ce356460e05be48d66401c1ae64ef84d163dd689964962c6888a9810865e39097a5e9de748876c2e0bf89b232d583c33982773e9903ae7a76257270986538",
+  "publishConfig": {
+    "access": "public",
+    "provenance": true
+  },
+  "keywords": [
+    "microfrontend",
+    "frontend",
+    "event",
+    "event bus",
+    "pubsub",
+    "subscription",
+    "publish",
+    "typescript",
+    "javascript",
+    "library"
+  ]
+}
