+++ packages/core/src/internal/event/navette-event.test.ts
