+++ packages/core/src/internal/json/json-compatible.test.ts
