+++ packages/core/src/internal/json/json-compatible.ts
