+++ packages/core/src/internal/topic/topic-config.ts
