+++ packages/core/src/internal/topic/topic-factory.ts
