+++ packages/core/src/internal/topic/topic.ts
