+++ packages/core/src/internal/topics-manager.ts
