+++ packages/core/src/main.ts
