+++ packages/core/stryker.config.js
