+++ packages/core/vite.config.js
